# apps/api/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/teachers/signup": {
      "get": {
        "operationId": "teacherSignup",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TeacherSignupDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } }
      }
    },
    "/teachers/login": {
      "post": {
        "operationId": "teacherLogin",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TeacherLoginDto" }
            }
          }
        },
        "responses": { "200": { "description": "teacher logged in" } }
      }
    },
...
    "/teachers/profile": {
      "get": {
        "operationId": "TeachersController_getProfile",
        "parameters": [],
        "responses": { "200": { "description": "" } }
      },
      "post": {
        "operationId": "TeachersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeacherUpdateProfileDto"
              }
            }
          }
        },
        "responses": { "201": { "description": "" } }
      }
    },
    "/teachers/questions": {
      "get": {
        "operationId": "TeachersController_getQuestions",
        "parameters": [],
        "responses": { "200": { "description": "" } }
      },
      "post": {
        "operationId": "TeachersController_createQuestion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QuestionCreateDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } }
      }
    },
    "/teachers/questions/{questionId}": {
      "get": {
        "operationId": "TeachersController_getQuestion",
        "parameters": [
          {
            "name": "questionId",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": { "200": { "description": "" } }
      }
    },
    "/teachers/{teacherId}/questions": {
      "get": {
        "operationId": "TeachersController_getParticularTeacherQuestions",
        "parameters": [
          {
            "name": "teacherId",
            "required": true,
            "in": "path",
            "schema": { "type": "number" }
          }
        ],
        "responses": { "200": { "description": "" } }
      }
    }
  },
  "info": {
    "title": "Questionnaire Open Api",
    "description": "Questionnaire api description",
    "version": "3.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "TeacherSignupDto": {
        "type": "object",
        "properties": {
          "firstname": { "type": "string" },
          "lastname": { "type": "string" },
          "username": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["firstname", "lastname", "username", "password"]
      },
      "TeacherLoginDto": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["username", "password"]
      },
      "TeacherUpdateProfileDto": {
        "type": "object",
        "properties": {
          "username": { "type": "string" },
          "firstname": { "type": "string" },
          "lastname": { "type": "string" },
          "password": { "type": "string" }
        },
        "required": ["username", "firstname", "lastname", "password"]
      },
      "QuestionCreateDto": { "type": "object", "properties": {} }
    }
  }
}